Rename django_sendfile.sendfile module to utils
Something that I missed when I renamed the package: the was a name
conflict between the imported sendfile function and the module it came
from as the former was imported directly into __init__.py
